requirements.txt
absl-py==1.0.0
aiohttp==3.5.4
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
atari-py==0.2.6
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
beautifulsoup4==4.12.2
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
cachetools==3.0.0
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.3.9
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.1
dask==1.0.0
decorator==4.3.0
deepdiff==3.3.0
distro==1.5.0
docutils==0.14
fasteners==0.15
filelock==3.0.12
flatbuffers==2.0
funcsigs==1.0.2
future==0.17.1
gast==0.2.0
gdown==4.7.1
gitdb2==2.0.5
GitPython==2.1.11
glfw==1.7.0
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-compute-engine==2.8.13
grpcio==1.16.1
gtimer==1.0.0b5
gym==0.15.4
h5py==2.8.0
git+https://github.com/minttusofia/hold-rewards.git@tf1_compat#egg=scenic
httplib2==0.12.0
hydra-core==1.3.2
idna==2.7
idna-ssl==1.1.0
imageio==2.4.1
importlib-metadata==1.6.0
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.17.0
jmespath==0.9.3
jsonpickle==1.4.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lockfile==0.12.2
lz4==3.0.2
Markdown==3.0.1
matplotlib==3.0.2
-e git+https://github.com/kschmeckpeper/metaworld.git@1d0d2c4d965b64b97723cadfeb635744968593e5#egg=metaworld
mock==4.0.2
monotonic==1.5
more-itertools==4.3.0
multidict==4.7.5
-e git+https://github.com/minttusofia/multiworld.git@55ca5b19736f2542080a8e15eddb75b1ce2e6d3b#egg=multiworld
mujoco-py<2.2,>=2.1
networkx==2.2
numpy==1.19.0
opencv-python
omegaconf==2.3.0
packaging==23.1
pandas==0.23.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
plotly==1.9.6
pluggy==0.8.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
py-spy==0.3.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycosat==0.6.3
pycparser==2.19
pyglet==1.3.2
Pygments==2.6.1
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyquaternion==0.9.5
PySocks==1.6.8
pytest==4.0.1
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==5.1.0
-e git+https://github.com/facebookresearch/r3m.git#egg=r3m
ray==0.6.6
redis==3.0.1
requests==2.20.1
rsa==3.4.2
ruamel.yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.14.2
scikit-learn==0.20.1
scipy==1.7.3
git+https://github.com/hartikainen/serializable.git@76516385a3a716ed4a2a9ad877e2d5cbcf18d4e6#egg=serializable
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.5
soupsieve==2.4.1
torch==1.12.1
torchvision==0.13.1
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-probability==0.6.0
termcolor==1.1.0
toolz==0.9.0
tqdm==4.47.0
traitlets==4.3.3
typing==3.7.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.9
Werkzeug==0.14.1
yarl==1.4.2
zipp==3.1.0
